# desc/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fishing Notes Admin API",
    "version": "0.2.1",
    "description": "管理后台 API 文档，覆盖认证、RBAC、管理员用户、钓点审核、知识专栏、鱼种内容与业务用户管理。\n\n使用说明：\n1. 除登录接口外，其余接口默认都需要 Bearer Token。\n2. 所有成功响应统一为 `{ code, message, data }`，其中 `code = 0` 表示成功。\n3. 列表接口统一支持 `page` 和 `pageSize`，默认 `page=1`、`pageSize=20`。\n"
  },
  "servers": [
    {
      "url": "http://localhost:1002",
      "description": "Local API Server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": [

      ]
    }
  ],
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Admin Users"
        ],
        "summary": "获取管理员用户列表",
        "operationId": "listAdminUsers",
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "username",
            "in": "query",
            "description": "按用户名模糊查询",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nickname",
            "in": "query",
            "description": "按昵称模糊查询",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "用户状态，字符串形式传入，例如 1 或 0",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAdminUsersEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Admin Users"
        ],
        "summary": "创建管理员账号",
        "operationId": "createAdminUser",
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/admin/users/{userId}": {
      "put": {
        "tags": [
          "Admin Users"
        ],
        "summary": "编辑管理员账号",
        "operationId": "updateAdminUser",
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/admin/users/{userId}/password": {
      "post": {
        "tags": [
          "Admin Users"
        ],
        "summary": "重置管理员密码",
        "operationId": "resetAdminPassword",
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetAdminPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重置成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/v1/banners": {
      "get": {
        "tags": ["Discover Banners"],
        "summary": "查询发现页横幅",
        "operationId": "listBanners",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/PageSize"},
          {"name": "keyword", "in": "query", "schema": {"type": "string"}},
          {"name": "visibleStatus", "in": "query", "schema": {"type": "string", "enum": ["0", "1"]}}
        ],
        "responses": {
          "200": {"description": "查询成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListBannersEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      },
      "post": {
        "tags": ["Discover Banners"],
        "summary": "创建发现页横幅",
        "operationId": "createBanner",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BannerUpsertRequest"}}}},
        "responses": {
          "200": {"description": "创建成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BannerEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/banners/{bannerId}": {
      "get": {
        "tags": ["Discover Banners"],
        "summary": "查询发现页横幅详情",
        "operationId": "getBanner",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/BannerIdPath"}],
        "responses": {
          "200": {"description": "查询成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BannerEnvelope"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      },
      "put": {
        "tags": ["Discover Banners"],
        "summary": "编辑发现页横幅",
        "operationId": "updateBanner",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/BannerIdPath"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BannerUpsertRequest"}}}},
        "responses": {
          "200": {"description": "更新成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BannerEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/v1/banners/{bannerId}/visibility": {
      "post": {
        "tags": ["Discover Banners"],
        "summary": "发布或隐藏发现页横幅",
        "operationId": "updateBannerVisibility",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/BannerIdPath"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateBannerVisibilityRequest"}}}},
        "responses": {
          "200": {"description": "操作成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuccessEnvelope"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "获取业务用户列表",
        "operationId": "listUsers",
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "按用户昵称或用户 ID 模糊查询",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "用户状态，字符串形式传入，例如 0、1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "所在城市精确筛选",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "description": "页码，默认 1",
        "schema": {
          "type": "integer",
          "format": "int64",
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "description": "每页数量，默认 20，最大 100",
        "schema": {
          "type": "integer",
          "format": "int64",
          "default": 20,
          "maximum": 100
        }
      },
      "UserIdPath": {
        "name": "userId",
        "in": "path",
        "required": true,
        "description": "用户 ID",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
      "BannerIdPath": {
        "name": "bannerId",
        "in": "path",
        "required": true,
        "description": "发现页横幅 ID",
        "schema": {"type": "integer", "format": "int64"}
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未认证或认证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "没有权限访问",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "服务端异常",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "RoleItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "type": "integer",
            "format": "int64"
          },
          "permissionCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateAdminUserRequest": {
        "type": "object",
        "required": [
          "username",
          "password",
          "nickname"
        ],
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 72
          },
          "nickname": {
            "type": "string",
            "maxLength": 64
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "UpdateAdminUserRequest": {
        "type": "object",
        "required": [
          "nickname",
          "status"
        ],
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 64
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          }
        }
      },
      "ResetAdminPasswordRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 72
          }
        }
      },
      "AdminUserItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "isSuper": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "format": "int64"
          },
          "lastLoginAt": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleItem"
            }
          }
        }
      },
      "ListAdminUsersResponse": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "BannerUpsertRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "tag": {"type": "string", "maxLength": 64},
          "title": {"type": "string", "maxLength": 128},
          "description": {"type": "string", "maxLength": 500},
          "deadlineText": {"type": "string", "maxLength": 128},
          "deadlineAt": {"type": "string", "format": "date"},
          "locationName": {"type": "string", "maxLength": 128},
          "sortOrder": {"type": "integer", "format": "int64"},
          "visibleStatus": {"type": "integer", "enum": [0, 1]}
        }
      },
      "UpdateBannerVisibilityRequest": {
        "type": "object",
        "required": ["visibleStatus"],
        "properties": {"visibleStatus": {"type": "integer", "enum": [0, 1]}}
      },
      "BannerItem": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "bannerCode": {"type": "string"},
          "tag": {"type": "string"},
          "title": {"type": "string"},
          "description": {"type": "string"},
          "deadlineText": {"type": "string"},
          "deadlineAt": {"type": "string", "format": "date"},
          "locationName": {"type": "string"},
          "sortOrder": {"type": "integer", "format": "int64"},
          "visibleStatus": {"type": "integer"},
          "publishedAt": {"type": "string", "format": "date-time"},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
      "ListBannersResponse": {
        "type": "object",
        "properties": {
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/BannerItem"}},
          "total": {"type": "integer", "format": "int64"},
          "page": {"type": "integer", "format": "int64"},
          "pageSize": {"type": "integer", "format": "int64"}
        }
      },
...
      "UserItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nickname": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int64"
          },
          "lastLoginAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "levelText": {
            "type": "string"
          },
          "profileDesc": {
            "type": "string"
          },
          "locationText": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "streakWeeks": {
            "type": "integer",
            "format": "int64"
          },
          "preferredSpeciesText": {
            "type": "string"
          }
        }
      },
      "ListUsersResponse": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SuccessData": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "example": {
          "success": true
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "example": {
          "code": 10000,
          "message": "invalid request"
        }
      },
...
      "ListAdminUsersEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/ListAdminUsersResponse"
          }
        }
      },
      "AdminUserEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserItem"
          }
        }
      },
...
      "ListBannersEnvelope": {
        "type": "object",
        "properties": {
          "code": {"type": "integer", "example": 0},
          "message": {"type": "string", "example": "ok"},
          "data": {"$ref": "#/components/schemas/ListBannersResponse"}
        }
      },
      "BannerEnvelope": {
        "type": "object",
        "properties": {
          "code": {"type": "integer", "example": 0},
          "message": {"type": "string", "example": "ok"},
          "data": {"$ref": "#/components/schemas/BannerItem"}
        }
      },
...
      "ListUsersEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/ListUsersResponse"
          }
        }
      },
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/SuccessData"
          }
        },
        "example": {
          "code": 0,
          "message": "ok",
          "data": {
            "success": true
          }
        }
      }
    }
  }
}